+++ R/step1_initialize.R
@@ -1,7 +1,43 @@
+
 #' Initialize absCopyNumber object from file or data.frame
 #'
-#' This is the first step for user to use a flexible way
+#' @description  This is the first step for user to use a flexible way for copynumber data input,
+#' absolute copy number (with purity and ploidy pair) calling. This function help
+#' user input their copy segmentation data as file or \code{data.frame}, input
+#' their SNV data as file or \code{data.frame}.
+#'
+#' @details support input format by absCopyNumber package please run \code{abs_supportfiles} function.
+#'
+#' @param seg The name of file or a \code{data.frame} containing the segmentation data.
+#' @param min.seg.len The minimum
+#' length of a segment to be included in computation. The default value is 200
+#' bp for WES, 3000 bp for WGS and 0 bp for MicroArray.
+#' @param snv optinal. The name of file or a \code{data.frame} containing a set of somatic single
+#' nucleotide variants (SNVs).
+#' @param isMaf logical. Whether the specified \code{snv} argument is provide as MAF file/data.frame.
+#' @param sample_seg character, default is \code{NULL}. At default function will look for 'sample' column in
+#' data (the best way to input multiple samples) of CNV or SNV. If 'sample' column is not find
+#' and \code{sample_seg} is specified,
+#' \code{sample_seg} will be treated as column name storing sample name or sample names (character vector),
+#' otherwise data will be treated as single sample data and sample name is 'sample'.
+#' @param sample_snv same as \code{sample_seg} but for snv data.
+#' @param platform character, default is 'WES'. Must be one of 'WES', 'WGS' or 'MicroArray'.
+#' @param verbose if \code{TRUE}, print extra information.
+#' @author Shixiang Wang <w_shixiang@163.com>
+#' @return a \code{absCopyNumber} object
 #' @import data.table methods
+#' @export
+#' @examples
+#' file_cn = system.file("extdata/example.cn.txt.gz", package = "absCopyNumber")
+#' file_snv = system.file("extdata/example.snv.txt.gz", package = "absCopyNumber")
+#'
+#' \donttest{
+#' res1 =  abs_initialize(seg = file_cn, snv = file_snv, verbose = T)
+#' res2 =  abs_initialize(seg = file_cn, snv = file_snv, sample_seg = "test1",
+#'                        sample_snv = "test1", verbose = T)
+#' }
+#'
+#' @seealso \code{\link[absCopyNumber]{absCopyNumber}}, \code{\link[absCopyNumber]{abs_prepare}}, , \code{\link[absCopyNumber]{abs_calling}}
 abs_initialize = function(
     seg,
     min.seg.len = NULL,
@@ -346,28 +382,7 @@ abs_initialize = function(
         }
     }
 
-    # absSummary = data.table::data.table(
-    #     nsample = length(unique(seg$sample)),
-    #     nchrom = length(unique(seg$chrom)),
-    #     max.seg = max(seg$eff.seg.len),
-    #     min.seg = min(seg$eff.seg.len),
-    #     max.ratio = max(seg$normalized.ratio),
-    #     min.ratio = min(seg$normalized.ratio)
-    # )
-
-    res = absCopyNumber(data = seg, SNV = snv, origin = origin)
-    # res = absCopyNumber(data = seg,
-    #                     snv.data = snv[if(exists("het.ind")) return(het.ind) else TRUE,],
-    #                     params=list(),
-    #                     origin=list(
-    #                         seg = seg_old,
-    #                         snv = snv
-    #                     ),
-    #                     summary=absSummary,
-    #                     result=data.table::data.table(),
-    #                     absCN=data.table::data.table(),
-    #                     absSNV=data.table::data.table())
-
+    res = absCopyNumber(data = seg, SNV = snv, origin = origin, params = list(platform=platform))
     if (verbose) {
         message("Done !")
     }
@@ -405,3 +420,23 @@ validata_snv = function(snv, verbose = FALSE) {
         stop("invalid snv input, please check!")
     }
 }
+
+
+# batch initializ : TODO?
+# abs_binitailize = function(batch_seg,
+#                            min.seg.len = NULL,
+#                            batch_snv = NULL,
+#                            isMaf = FALSE,
+#                            samplenames = NULL,
+#                            platform = c("WES", "WGS", "MicroArray")
+# ){
+#
+# }
+
+# Hack global variable  ---------------------------------------------------
+base::suppressMessages(utils::globalVariables(c("..cols", "Chromosome",
+                                                "Start", "End", "Segment_Mean",
+                                                "Variant_Type", "..maf_cols",
+                                                "Tumor_Sample_Barcode",
+                                                "Start_position", "t_alt_count",
+                                                "t_ref_count", "eff.seg.len")))