+++ NEWS.md
@@ -7,4 +7,9 @@
 * fix many doc problems and optimize output when verbose option set to `TRUE` (add TxT progress bar)
 
 
+# absCopyNumber 0.2.3
 
+* create a new pipeline to substitute old ways
+    * `abs_initialize` to load data
+    * `abs_prepare` to load parameters
+    * `abs_calling` to call results